// src/app/client/model/placeDto.ts
/**
 * con-api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PlaceType } from './placeType';
import { PrivacyLevel } from './privacyLevel';


/**
 * DTO for place creation/update
 */
export interface PlaceDto { 
    id?: string | null;
    name?: string | null;
    type?: PlaceType;
    parentPlaceId?: string | null;
    latitude?: number | null;
    longitude?: number | null;
    privacyLevel?: PrivacyLevel;
}
export namespace PlaceDto {
}