// src/app/client/model/place.ts
/**
 * con-api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PlaceType } from './placeType';
import { PrivacyLevel } from './privacyLevel';


/**
 * Place entity with hierarchical support
 */
export interface Place { 
    /**
     * Unique identifier for this entity
     */
    id?: string;
    /**
     * User who owns this entity
     */
    userId?: string;
    /**
     * When this entity was created
     */
    createdAt?: string;
    /**
     * When this entity was last modified
     */
    updatedAt?: string;
    privacyLevel?: PrivacyLevel;
    /**
     * Name of the place
     */
    name?: string | null;
    type?: PlaceType;
    /**
     * Parent place ID for hierarchy (e.g., village -> county -> country)
     */
    parentPlaceId?: string | null;
    /**
     * Latitude coordinate
     */
    latitude?: number | null;
    /**
     * Longitude coordinate
     */
    longitude?: number | null;
    /**
     * Full hierarchical path (e.g., \"Germany/Bavaria/Munich\")
     */
    hierarchyPath?: string | null;
    /**
     * Flexible attributes stored as a map of key -> value
     */
    attributes?: { [key: string]: string; } | null;
}
export namespace Place {
}